calc2.0

+++ Calc_01/Calc_01/Converter.cs
@@ -7,12 +7,12 @@ namespace Calc_01
 {
     public class Converter
     {
-        public static int from8to10(int oct)
+        public static Int64 from8to10(Int64 oct)
         {
-            int dec = 0;
-            int sign = (oct >= 0) ? +1 : -1;
+            Int64 dec = 0;
+            Int64 sign = (oct >= 0) ? +1 : -1;
             oct = Math.Abs(oct);
-            int k = 1;
+            Int64 k = 1;
             while (oct > 0)
             {
                 dec += (oct % 10) * k;
@@ -22,12 +22,12 @@ namespace Calc_01
             return dec * sign;
         }
 
-        public static int from10to8(int dec)
+        public static Int64 from10to8(Int64 dec)
         {
-            int oct = 0;
-            int sign = (dec >= 0) ? +1 : -1;
+            Int64 oct = 0;
+            Int64 sign = (dec >= 0) ? +1 : -1;
             dec = Math.Abs(dec);
-            int k = 1;
+            Int64 k = 1;
             while (dec > 0)
             {
                 oct += (dec % 8) * k;
